# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Ok": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "Ok",
        "type": "object"
      },
...
      "Profile": {
        "properties": {
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_demo": {
            "title": "Is Demo",
            "type": "boolean"
          },
          "locale": {
            "title": "Locale",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "week_start": {
            "title": "Week Start",
            "type": "integer"
          },
          "weight_unit": {
            "enum": [
              "kg",
              "lb"
            ],
            "title": "Weight Unit",
            "type": "string"
          }
        },
        "required": [
          "id",
          "version",
          "name",
          "timezone",
          "currency",
          "locale",
          "week_start",
          "weight_unit",
          "is_demo"
        ],
        "title": "Profile",
        "type": "object"
      },
      "ProfilePatch": {
        "additionalProperties": false,
        "properties": {
          "currency": {
            "anyOf": [
              {
                "pattern": "^[A-Z]{3}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "locale": {
            "anyOf": [
              {
                "enum": [
                  "pt-BR",
                  "en-US"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locale"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "timezone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timezone"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "week_start": {
            "anyOf": [
              {
                "maximum": 6.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Week Start"
          },
          "weight_unit": {
            "anyOf": [
              {
                "enum": [
                  "kg",
                  "lb"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Unit"
          }
        },
        "required": [
          "version"
        ],
        "title": "ProfilePatch",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Orbit API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/demo/reset": {
      "post": {
        "operationId": "reset_api_v1_auth_demo_reset_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reset"
      }
    },
...
    "/api/v1/me": {
      "get": {
        "operationId": "me_api_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Me"
      },
      "patch": {
        "operationId": "edit_me_api_v1_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Profile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Edit Me"
      }
    },
...
  }
}